% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{assign_colors}
\alias{assign_colors}
\title{assign colors by a column in a tibble, for the convenience to
use \code{scale_color_identity()}}
\usage{
assign_colors(df, by, colors = sci_colors("npg", 10), na = "#F5F5F5")
}
\arguments{
\item{df}{tibble}

\item{by}{assign colors according to this column}

\item{colors}{a vector of color values}

\item{na}{if colors are not enough, fill na values}
}
\value{
tibble
}
\description{
assign colors by a column in a tibble, for the convenience to
use \code{scale_color_identity()}
}
\examples{
assign_colors(mini_diamond, cut, colors = sci_colors("nejm", 8))
}
