% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bioletter_colors}
\alias{bioletter_colors}
\title{colors of nucleotides and amino acids}
\format{
\subsection{\code{bioletter_colors}}{

colors for biological letters like amino acids or nucleotides
}
}
\source{
according to the print format
}
\usage{
bioletter_colors
}
\description{
colors of nucleotides and amino acids
}
\keyword{datasets}
