% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{brewer_colors}
\alias{brewer_colors}
\title{select colors from \code{RColorBrewer} package presets}
\usage{
brewer_colors(name, n = 3, ...)
}
\arguments{
\item{name}{presets name}

\item{n}{number of colors}

\item{...}{other arguments of \code{RColorBrewer::brewer.pal}}
}
\value{
colors
}
\description{
select colors from \code{RColorBrewer} package presets
}
\examples{
brewer_colors("Blues", 5)
}
