% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canvas_size}
\alias{canvas_size}
\title{width and height of built-in canvas}
\format{
\subsection{\code{canvas_size}}{

canvas sizes list
}
}
\source{
according to the print format
}
\usage{
canvas_size
}
\description{
width and height of built-in canvas
}
\keyword{datasets}
