% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-CommodityFuturesAndOptions.R
\docType{class}
\name{CommodityFuturesAndOptions-class}
\alias{CommodityFuturesAndOptions-class}
\alias{CommodityFuturesAndOptions}
\title{Commodity Futures and Options}
\description{
Commodity Futures and Options
}
\section{Methods}{

\describe{
\item{\code{CmeEod()}}{Query the end-of-day prices of commodity futures contracts traded on the COMEX and NYMEX}

\item{\code{McxEod()}}{Query the end-of-day prices of commodity contracts traded on the MCX}

\item{\code{NcdexEod()}}{Query the end-of-day prices of commodity contracts traded on the NCDEX}
}}

\references{
\url{https://www.mcxindia.com/}
\url{https://www.ncdex.com/}
\url{https://www.cmegroup.com/}

Read the python documentation for information on the data-attributes \url{https://plutopy.readthedocs.io/en/latest/CommodityFuturesAndOptions.html}

Sample notebook: \url{https://github.com/shyams80/plutons/blob/master/docs-R/CommodityFuturesAndOptions.ipynb}
}
