% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-Currencies.R
\docType{class}
\name{Currencies-class}
\alias{Currencies-class}
\alias{Currencies}
\title{Currencies}
\description{
Currencies
}
\section{Methods}{

\describe{
\item{\code{Av30minTimeSeries()}}{Query the 30-min bar of currency USD-fx time-series published by AlphaVantage}

\item{\code{AvEodTimeSeries()}}{Query the end-of-day currency USD-fx time-series published by AlphaVantage}

\item{\code{NseFuturesTimeSeries()}}{Query the currency futures time-series published by the NSE}

\item{\code{NseOptionsTimeSeries()}}{Query the currency option time-series published by the NSE}
}}

\references{
\url{https://nseindia.com/}
\url{https://www.alphavantage.co}

Read the python documentation for information on the data-attributes \url{https://plutopy.readthedocs.io/en/latest/Currencies.html}

Sample notebook: \url{https://github.com/shyams80/plutons/blob/master/docs-R/Currencies.ipynb}
}
