% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-EquitiesFuturesAndOptionsIndiaNse.R
\docType{class}
\name{EquitiesFuturesAndOptionsIndiaNse-class}
\alias{EquitiesFuturesAndOptionsIndiaNse-class}
\alias{EquitiesFuturesAndOptionsIndiaNse}
\title{Equities Futures and Options traded on the NSE}
\description{
Equities Futures and Options traded on the NSE
}
\section{Methods}{

\describe{
\item{\code{FuturesEodTimeSeries()}}{Query the end-of-day prices of equity futures contracts traded on the NSE}

\item{\code{LotSize()}}{Query the lot-size of equity futures and options contracts traded on the NSE}

\item{\code{OptionGreeks()}}{Query the end-of-day option greeks of equity options contracts traded on the NSE}

\item{\code{OptionsEodTimeSeries()}}{Query the end-of-day prices of equity options contracts traded on the NSE}
}}

\references{
\url{https://nseindia.com/}

Read the python documentation for information on the data-attributes \url{https://plutopy.readthedocs.io/en/latest/EquitiesFuturesAndOptionsIndiaNse.html}

Sample notebook: \url{https://github.com/shyams80/plutons/blob/master/docs-R/EquitiesFuturesAndOptionsIndiaNse.ipynb}
}
