% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-EquitiesIndiaBse.R
\docType{class}
\name{EquitiesIndiaBse-class}
\alias{EquitiesIndiaBse-class}
\alias{EquitiesIndiaBse}
\title{Equities India BSE}
\description{
Equities India BSE
}
\section{Methods}{

\describe{
\item{\code{CorporateResults()}}{Query to CorporateResults for a specific period and stock}

\item{\code{EodTimeSeries()}}{Query unadjusted end-of-day price and volume time-series for listed stocks}

\item{\code{MiscInfo()}}{Query miscellaneous information of listed equity tickers}

\item{\code{Tickers()}}{Query meta data of listed equity tickers}
}}

\references{
\url{https://bseindia.com/}

Read the python documentation for information on the data-attributes \url{https://plutopy.readthedocs.io/en/latest/EquitiesIndiaBse.html}

Sample notebook: \url{https://github.com/shyams80/plutons/blob/master/docs-R/EquitiesIndiaBse.ipynb}
}
