% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-EquitiesUs.R
\docType{class}
\name{EquitiesUs-class}
\alias{EquitiesUs-class}
\alias{EquitiesUs}
\title{Equities US}
\description{
Equities US
}
\section{Methods}{

\describe{
\item{\code{EodAdjustedTimeSeries()}}{Query end-of-day price and volume time-series adjusted for splits, bonus and dividends for listed stocks}

\item{\code{SecFilings()}}{Links to SEC filings. Useful to keep track of M&A events in your portfolio}

\item{\code{SecMeta()}}{Meta information of companies in the SEC EDGAR database. Useful for SIC.}

\item{\code{Tickers()}}{Query meta data of listed equity (common stock) tickers}
}}

\references{
\url{https://iexcloud.io}
\url{https://www.sec.gov/}

Read the python documentation for information on the data-attributes \url{https://plutopy.readthedocs.io/en/latest/EquitiesUs.html}

Sample notebook: \url{https://github.com/shyams80/plutons/blob/master/docs-R/EquitiesUs.ipynb}
}
