% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-ETFsUs.R
\docType{class}
\name{ETFsUs-class}
\alias{ETFsUs-class}
\alias{ETFsUs}
\title{ETFs listed in the US}
\description{
Read the python documentation for information on the data-attributes \url{https://plutopy.readthedocs.io/en/latest/ETFsUs.html}
}
\details{
Sample notebook: \url{https://github.com/shyams80/plutons/blob/master/docs-R/ETFsUs.ipynb}
}
\section{Methods}{

\describe{
\item{\code{Meta()}}{Query the meta-data}
}}

