% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_compare.R
\name{extract_compare}
\alias{extract_compare}
\title{extract the result of \code{geom_compare}}
\usage{
extract_compare(p)
}
\arguments{
\item{p}{ggplot object}
}
\value{
compare tibble
}
\description{
extract the result of \code{geom_compare}
}
\examples{

p <- ggplot(data = mini_diamond, mapping = aes(x = cut, y = price)) +
  geom_point() +
  geom_describe(show_error = FALSE, color = "red") +
  geom_compare(
    cp_label = c("psymbol", "right_deno_fc"),
    lab_pos = 20000, step_increase = 0.22
  ) +
  ylim(0, 30000)

extract_compare(p)

}
