% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-FamaFrench.R
\docType{class}
\name{FamaFrench-class}
\alias{FamaFrench-class}
\alias{FamaFrench}
\title{FamaFrench}
\description{
FamaFrench
}
\section{Methods}{

\describe{
\item{\code{FiveFactor3x2Daily()}}{Query the Fama-French 5-factor daily returns (http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/Data_Library/f-f_5_factors_2x3.html)}

\item{\code{Industry49Daily()}}{Query the Fama-French daily returns of 49 different industries (http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/Data_Library/det_49_ind_port.html)}

\item{\code{MomentumDaily()}}{Query the Fama-French daily returns of momentum factor and portfolios}

\item{\code{MomentumMonthly()}}{Query the Fama-French monthly returns of momentum factor and portfolios}
}}

\references{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/index.html}

Read the python documentation for information on the data-attributes \url{https://plutopy.readthedocs.io/en/latest/FamaFrench.html}

Sample notebook: \url{https://github.com/shyams80/plutons/blob/master/docs-R/FamaFrench.ipynb}
}
