% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-Fred.R
\docType{class}
\name{Fred-class}
\alias{Fred-class}
\alias{Fred}
\title{St. Louis Fed FRED cache}
\description{
St. Louis Fed FRED cache
}
\section{Methods}{

\describe{
\item{\code{Meta()}}{Query the FRED meta-data cache}

\item{\code{TimeSeries()}}{Query the FRED time-series cache}
}}

\references{
\url{https://fred.stlouisfed.org/}

Read the python documentation for information on the data-attributes \url{https://plutopy.readthedocs.io/en/latest/Fred.html}

Sample notebook: \url{https://github.com/shyams80/plutons/blob/master/docs-R/Fred.ipynb}
}
