% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_compare.R
\name{geom2trace.GeomCompare}
\alias{geom2trace.GeomCompare}
\title{geom2trace.GeomCompare}
\usage{
geom2trace.GeomCompare(data, params, plot)
}
\arguments{
\item{data, params, plot}{params}
}
\value{
no return value
}
\description{
geom2trace.GeomCompare
}
