% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_describe.R
\name{geom2trace.GeomDescribe}
\alias{geom2trace.GeomDescribe}
\title{geom2trace.GeomDescribe}
\usage{
geom2trace.GeomDescribe(data, params, plot)
}
\arguments{
\item{data, params, plot}{params}
}
\value{
no return value
}
\description{
geom2trace.GeomDescribe
}
