% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_describe.R
\name{geom_describe}
\alias{geom_describe}
\title{Description values plot}
\usage{
geom_describe(
  mapping = NULL,
  data = NULL,
  stat = "describe",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  lineend = "round",
  show_error = TRUE,
  center_symbol = "bar",
  center_width = 0.3,
  error_width = 0.2,
  center_func = mean,
  low_func = function(x, na.rm) {
     mean(x, na.rm = na.rm) - sd(x, na.rm = na.rm)
 },
  high_func = function(x, na.rm) {
     mean(x, na.rm = na.rm) + sd(x, na.rm = na.rm)
 },
  ...
)
}
\arguments{
\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show_error}{show error symbol}

\item{center_symbol}{one of \verb{point, bar}}

\item{center_width}{if \code{center_symbol='bar'}, the width of the bar}

\item{error_width}{the width of the error bar}

\item{center_func}{the center function, \code{mean} as default}

\item{low_func}{the low error function, \code{mean} minus \code{sd} as default}

\item{high_func}{the high error function, \code{mean} plus \code{sd} as default}

\item{...}{Other arguments passed on to \code{layer()}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
The describe geom is used to create description values plot, including
center symbol and error symbol. The center symbol can be mean, median or
other custom functions, the error symbol can be sd, quantile or other custom
functions.
}
\examples{

mini_diamond \%>\% ggplot(aes(x = cut, y = price)) +
  geom_point() +
  geom_describe()

}
