% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_compare.R
\docType{data}
\name{GeomCompare}
\alias{GeomCompare}
\title{GeomCompare}
\format{
An object of class \code{GeomCompare} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomCompare
}
\description{
GeomCompare
}
\keyword{datasets}
