% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_describe.R
\docType{data}
\name{GeomDescribe}
\alias{GeomDescribe}
\title{GeomDescribe}
\format{
An object of class \code{GeomDescribe} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomDescribe
}
\description{
GeomDescribe
}
\keyword{datasets}
