% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{gradient_colors}
\alias{gradient_colors}
\title{generate gradient colors}
\usage{
gradient_colors(x, n)
}
\arguments{
\item{x}{colors}

\item{n}{number of colors to output}
}
\value{
gradient colors
}
\description{
generate gradient colors
}
\examples{
gradient_colors(c("blue", "red"), 10)
}
