% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{inch2cm}
\alias{inch2cm}
\alias{in2cm}
\alias{cm2in}
\title{trans inch to cm}
\usage{
inch2cm(x)

in2cm(x)

cm2in(x)
}
\arguments{
\item{x}{inch value}
}
\value{
cm value
}
\description{
trans inch to cm
}
\examples{
inch2cm(1)
}
