% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-Indices.R
\docType{class}
\name{Indices-class}
\alias{Indices-class}
\alias{Indices}
\title{Indices}
\description{
Indices
}
\section{Methods}{

\describe{
\item{\code{BseConstituents()}}{Query the latest constituents of BSE indices}

\item{\code{BseTimeSeries()}}{Query the index time-series published by the BSE}

\item{\code{FamaFrench5Factor3x2Daily()}}{Query the Fama-French 5-factor daily returns (http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/Data_Library/f-f_5_factors_2x3.html)}

\item{\code{FamaFrenchIndustry49Daily()}}{Query the Fama-French daily returns of 49 different industries (http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/Data_Library/det_49_ind_port.html)}

\item{\code{IndiaGsecTimeSeries()}}{Query the Indian Government Soverign Bond index time-series published by the CCIL}

\item{\code{IndiaVixTimeSeries()}}{Query the India VIX time-series data published by the NSE}

\item{\code{NseConstituents()}}{Query the latest constituents of NSE indices}

\item{\code{NseTimeSeries()}}{Query the index time-series published by the NSE}

\item{\code{YahooFinanceTimeSeries()}}{Query the index time-series published by Yahoo Finance}
}}

\references{
\url{https://nseindia.com/}
\url{https://bseindia.com/}
\url{https://www.ccilindia.com}
\url{https://finance.yahoo.com/}
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/index.html}

Read the python documentation for information on the data-attributes \url{https://plutopy.readthedocs.io/en/latest/Indices.html}

Sample notebook: \url{https://github.com/shyams80/plutons/blob/master/docs-R/Indices.ipynb}
}
