% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-InternationalMonetaryFund.R
\docType{class}
\name{InternationalMonetaryFund-class}
\alias{InternationalMonetaryFund-class}
\alias{InternationalMonetaryFund}
\title{International Monetary Fund}
\description{
International Monetary Fund
}
\section{Methods}{

\describe{
\item{\code{Meta()}}{Query IMF meta-data based on geography/indicator. Then use the ID obtained here to query the time-series}

\item{\code{TimeSeries()}}{Query the IMF time-series data based on the Meta ID}
}}

\references{
\url{https://www.imf.org}

Read the python documentation for information on the data-attributes \url{https://plutopy.readthedocs.io/en/latest/InternationalMonetaryFund.html}

Sample notebook: \url{https://github.com/shyams80/plutons/blob/master/docs-R/InternationalMonetaryFund.ipynb}
}
