% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-InvestmentFlowsIndia.R
\docType{class}
\name{InvestmentFlowsIndia-class}
\alias{InvestmentFlowsIndia-class}
\alias{InvestmentFlowsIndia}
\title{Investment Flows India}
\description{
Investment Flows India
}
\section{Methods}{

\describe{
\item{\code{DiiCashMarket()}}{Query purchases and sales of debt and equity by Domestic Institutional Investors}

\item{\code{DiiDerivativesMarket()}}{Query purchases and sales of futures and options by Domestic Institutional Investors}

\item{\code{FiiCashMarket()}}{Query purchases and sales of debt and equity by Foreign Institutional Investors}

\item{\code{FiiDerivativesMarket()}}{Query purchases and sales of futures and options by Foreign Institutional Investors}
}}

\references{
\url{https://cdslindia.com/}
\url{https://www.sebi.gov.in}

Read the python documentation for information on the data-attributes \url{https://plutopy.readthedocs.io/en/latest/InvestmentFlowsIndia.html}

Sample notebook: \url{https://github.com/shyams80/plutons/blob/master/docs-R/InvestmentFlowsIndia.ipynb}
}
