% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{lpt}
\alias{lpt}
\title{trans geom line point and theme line point to the real point}
\usage{
lpt(x)
}
\arguments{
\item{x}{line point in geom or theme}
}
\value{
real point
}
\description{
trans geom line point and theme line point to the real point
}
\examples{
lpt(1)
}
