% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_diamond}
\alias{mini_diamond}
\title{base color of munsell}
\format{
\subsection{\code{base_color}}{
}

\subsection{\code{canvas_size}}{
}
}
\source{
adjusted from \code{munsell} package

width and height of built-in canvas

built according to the web
}
\usage{
mini_diamond
}
\description{
base color of munsell
}
\keyword{datasets}
