% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-MutualFundsIndia.R
\docType{class}
\name{MutualFundsIndia-class}
\alias{MutualFundsIndia-class}
\alias{MutualFundsIndia}
\title{Amfi India mutual-fund cache}
\description{
Amfi India mutual-fund cache
}
\section{Methods}{

\describe{
\item{\code{AumFundwise()}}{Query the fund-wise assets under management (AUM) of different asset managers}

\item{\code{AumSchemewise()}}{Query the scheme-wise assets under management (AUM) of different asset managers and their 'schemes'}

\item{\code{Meta()}}{Query the meta-data}

\item{\code{NavTimeSeries()}}{Query for the NAVs of different funds}

\item{\code{Portfolio()}}{Query for the monthly reported portfolios of different funds}
}}

\references{
\url{https://amfiindia.com/}

Read the python documentation for information on the data-attributes \url{https://plutopy.readthedocs.io/en/latest/MutualFundsIndia.html}

Sample notebook: \url{}
}
