% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-NasdaqOmx.R
\docType{class}
\name{NasdaqOmx-class}
\alias{NasdaqOmx-class}
\alias{NasdaqOmx}
\title{NasdaqOmx Indices}
\description{
NasdaqOmx Indices
}
\section{Methods}{

\describe{
\item{\code{Meta()}}{Query to find the time-series id of total-return indices published by the NasdaqOmx}

\item{\code{TimeSeries()}}{Query the index time-series published by the NasdaqOmx}
}}

\references{
\url{https://indexes.nasdaqomx.com/}
\url{https://www.quandl.com/}

Read the python documentation for information on the data-attributes \url{https://plutopy.readthedocs.io/en/latest/NasdaqOmx.html}

Sample notebook: \url{https://github.com/shyams80/plutons/blob/master/docs-R/NasdaqOmx.ipynb}
}
