% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarts-theme.R
\name{pe_color}
\alias{pe_color}
\title{set colors for echarts plot}
\usage{
pe_color(e, x, rep = FALSE, default = "transparent")
}
\arguments{
\item{e}{echarts4r object}

\item{x}{color character vector}

\item{rep}{replicate the colors if needed}

\item{default}{default color}
}
\description{
set colors for echarts plot
}
\examples{

mini_diamond \%>\%
  dplyr::group_by(cut) \%>\%
  e_charts(x, width = 500, height = 500) \%>\%
  e_scatter(y, symbol_size = 10) \%>\%
  pe_color(c("red", "yellow"), default = "black")

}
