% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{pl_init}
\alias{pl_init}
\title{set pl_size of the}
\usage{
pl_init(
  width = 3,
  height = 1.5,
  res = 300,
  w = NULL,
  h = NULL,
  theme = theme_pl()
)
}
\arguments{
\item{width}{width}

\item{height}{height}

\item{res}{resolution, 300 as default}

\item{w}{alias of width}

\item{h}{alias of height}

\item{theme}{default theme}
}
\description{
set pl_size of the
}
\examples{
pl_init()
}
