% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{pl_save}
\alias{pl_save}
\title{save plot, support save into a blank canvas}
\usage{
pl_save(
  plot,
  filename,
  width,
  height,
  units = "in",
  canvas = NULL,
  canvas_pos_x = 0.5,
  canvas_pos_y = 0.1,
  ...
)
}
\arguments{
\item{plot}{ggplot object}

\item{filename}{filename}

\item{width}{plot width}

\item{height}{plot height}

\item{units}{units, 'in' for inch as default. Can be 'in', 'cm'}

\item{canvas}{\code{NULL} as default, pass character to use
built-in canvas ('A4', 'A4v'), or pass a numeric vector in 'c(width, heigh)'
form}

\item{canvas_pos_x}{from 0 to 1, the horizontal position of plot in canvas}

\item{canvas_pos_y}{from 0 to 1, the vertical position of plot in canvas}

\item{...}{other arguments from \code{ggsave}}
}
\value{
no return value
}
\description{
save plot, support save into a blank canvas
}
