% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{pl_size}
\alias{pl_size}
\title{set repr size and resolution}
\usage{
pl_size(width = 4, height = 3, res = 300, w = NULL, h = NULL)
}
\arguments{
\item{width}{width}

\item{height}{height}

\item{res}{resolution, 300 as default}

\item{w}{alias of width}

\item{h}{alias of height}
}
\value{
no return value
}
\description{
set repr size and resolution
}
\examples{
pl_size(width = 3, height = 2)
}
