% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{plot_colors}
\alias{plot_colors}
\title{show colors}
\usage{
plot_colors(x, ncol = 10, show_name = TRUE)
}
\arguments{
\item{x}{color values}

\item{ncol}{color number of each row}

\item{show_name}{use vector names as label}
}
\value{
ggplot object
}
\description{
show colors
}
\examples{
plot_colors(gradient_color(c("blue", "red"), 10))
}
