% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plutor-package.R
\docType{package}
\name{plutor-package}
\alias{plutor}
\alias{plutor-package}
\title{plutor: plot using ggplot}
\description{
plot using ggplot
}
\author{
\strong{Maintainer}: William Song \email{william_swl@163.com}

}
\keyword{internal}
