% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Position.R
\name{position_floatyPL}
\alias{position_floatyPL}
\title{a new Position object to create float y position}
\usage{
position_floatyPL(float = -0.05, cycle = 2)
}
\arguments{
\item{float}{float range, a ratio according to the whole panel height}

\item{cycle}{float cycle}
}
\value{
Position object
}
\description{
a new Position object to create float y position
}
