% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Position.R
\docType{data}
\name{PositionFloatxPL}
\alias{PositionFloatxPL}
\title{PositionFloatxPL}
\format{
An object of class \code{PositionFloatxPL} (inherits from \code{Position}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
PositionFloatxPL
}
\description{
PositionFloatxPL
}
\keyword{datasets}
