% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Position.R
\docType{data}
\name{PositionFloatyPL}
\alias{PositionFloatyPL}
\title{PositionFloatyPL}
\format{
An object of class \code{PositionFloatyPL} (inherits from \code{Position}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
PositionFloatyPL
}
\description{
PositionFloatyPL
}
\keyword{datasets}
