% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{revert_pos_scale}
\alias{revert_pos_scale}
\title{revert the position scale transformation}
\usage{
revert_pos_scale(s)
}
\arguments{
\item{s}{ScaleContinuousPosition object, e.g. scales$y in compute_group()}
}
\value{
function
}
\description{
revert the position scale transformation
}
