% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_ele}
\alias{scale_ele}
\title{scale element according to a vector of element scales}
\usage{
scale_ele(level, base, ele_scales)
}
\arguments{
\item{level}{output level}

\item{base}{value of base level}

\item{ele_scales}{vector of element scales}
}
\value{
value of output level
}
\description{
scale element according to a vector of element scales
}
\examples{
scale_ele(level = 2, base = 5, ele_scales = c(1, 2))
}
