% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scales.R
\name{scale_y_continuous_pl}
\alias{scale_y_continuous_pl}
\title{A variant of \code{scale_y_continuous()} to show axis minor breaks}
\usage{
scale_y_continuous_pl(
  name = waiver(),
  breaks = waiver(),
  minor_breaks = NULL,
  n.breaks = NULL,
  labels = waiver(),
  limits = NULL,
  expand = ggplot2::expansion(),
  oob = scales::oob_keep,
  na.value = NA_real_,
  trans = "identity",
  guide = ggh4x::guide_axis_minor(),
  position = "left",
  sec.axis = waiver(),
  show_minor_breaks = TRUE,
  minor_break_step = NULL
)
}
\arguments{
\item{expand}{use \code{expansion()} to dismiss the blank between y axis low limit
and x axis}

\item{oob}{use \code{scales::oob_keep} instead of \code{scales::oob_censor}, which
will always consider the data points out of the limits}

\item{show_minor_breaks}{show minor breaks or not}

\item{minor_break_step}{the step of minor breaks}
}
\description{
A variant of \code{scale_y_continuous()} to show axis minor breaks
}
