% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scales.R
\name{scale_y_log10_pl}
\alias{scale_y_log10_pl}
\title{A variant of \code{scale_y_log10()} to show axis minor breaks
and better axis labels}
\usage{
scale_y_log10_pl(
  name = waiver(),
  breaks = NULL,
  minor_breaks = NULL,
  n.breaks = NULL,
  labels = NULL,
  limits = NULL,
  expand = ggplot2::expansion(),
  oob = scales::oob_keep,
  na.value = NA_real_,
  trans = scales::log10_trans(),
  guide = ggh4x::guide_axis_minor(),
  position = "left",
  sec.axis = waiver(),
  show_minor_breaks = TRUE
)
}
\arguments{
\item{expand}{use \code{expansion()} to dismiss the blank between y axis low limit
and x axis}

\item{oob}{use \code{scales::oob_keep} instead of \code{scales::oob_censor}, which
will always consider the data points out of the limits}

\item{show_minor_breaks}{show minor breaks or not}
}
\description{
A variant of \code{scale_y_log10()} to show axis minor breaks
and better axis labels
}
