% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{sci_colors}
\alias{sci_colors}
\title{select colors from \code{ggsci} package presets}
\usage{
sci_colors(name, n = 3, alpha = 1)
}
\arguments{
\item{name}{presets name}

\item{n}{number of colors}

\item{alpha}{alpha}
}
\value{
colors
}
\description{
select colors from \code{ggsci} package presets
}
\examples{
sci_colors("npg", 5)
}
