% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_compare.R
\docType{data}
\name{StatCompare}
\alias{StatCompare}
\title{StatCompare}
\format{
An object of class \code{StatCompare} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
StatCompare
}
\description{
StatCompare
}
\keyword{datasets}
