% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\docType{data}
\name{StatCountPL}
\alias{StatCountPL}
\title{StatCountPL}
\format{
An object of class \code{StatCountPL} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
StatCountPL
}
\description{
StatCountPL
}
\keyword{datasets}
