% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_describe.R
\docType{data}
\name{StatDescribe}
\alias{StatDescribe}
\title{StatDescribe}
\format{
An object of class \code{StatDescribe} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
StatDescribe
}
\description{
StatDescribe
}
\keyword{datasets}
