% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\docType{data}
\name{StatFuncPL}
\alias{StatFuncPL}
\title{StatFuncPL}
\format{
An object of class \code{StatFuncPL} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
StatFuncPL
}
\description{
StatFuncPL
}
\keyword{datasets}
