% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\docType{data}
\name{StatMeanPL}
\alias{StatMeanPL}
\title{StatMeanPL}
\format{
An object of class \code{StatMeanPL} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
StatMeanPL
}
\description{
StatMeanPL
}
\keyword{datasets}
