% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_pl0}
\alias{theme_pl0}
\title{a blank theme}
\usage{
theme_pl0(...)
}
\arguments{
\item{...}{arguments of theme_pl()}
}
\value{
theme object of ggplot
}
\description{
a blank theme
}
\examples{
theme_pl0()
}
