% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{tpt}
\alias{tpt}
\title{trans geom text point to the real point}
\usage{
tpt(x)
}
\arguments{
\item{x}{text point in geom}
}
\value{
real point
}
\description{
trans geom text point to the real point
}
\examples{
tpt(5)
}
