% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{trans_pos_scale}
\alias{trans_pos_scale}
\title{perform the position scale transformation}
\usage{
trans_pos_scale(s)
}
\arguments{
\item{s}{ScaleContinuousPosition object, e.g. scales$y in compute_group()}
}
\value{
function
}
\description{
perform the position scale transformation
}
