% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-WorldBank.R
\docType{class}
\name{WorldBank-class}
\alias{WorldBank-class}
\alias{WorldBank}
\title{World Bank Global Economic Monitor data}
\description{
World Bank Global Economic Monitor data
}
\section{Methods}{

\describe{
\item{\code{Meta()}}{Query to find the country and indicator id's of the Global Economic Monitor data published by the World Bank}

\item{\code{TimeSeries()}}{Query the Global Economic Monitor data the World Bank}
}}

\references{
\url{https://worldbank.org}

Read the python documentation for information on the data-attributes \url{https://plutopy.readthedocs.io/en/latest/WorldBank.html}

Sample notebook: \url{https://github.com/shyams80/plutons/blob/master/docs-R/WorldBank.ipynb}
}
