% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-Yale.R
\docType{class}
\name{Yale-class}
\alias{Yale-class}
\alias{Yale}
\title{Yale/Shiller}
\description{
Yale/Shiller
}
\section{Methods}{

\describe{
\item{\code{Confidence()}}{Query Stock Market Confidence Indexes produced by the Yale School of Management. (https://svz.bz/2NzetU3)}

\item{\code{SP500()}}{Query the monthly stock price, dividends, and earnings data and the consumer price index.}
}}

\references{
\url{http://www.econ.yale.edu/~shiller/data.htm}

Read the python documentation for information on the data-attributes \url{https://plutopy.readthedocs.io/en/latest/Yale.html}

Sample notebook: \url{https://github.com/shyams80/plutons/blob/master/docs-R/Yale.ipynb}
}
