% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-YieldCurve.R
\docType{class}
\name{YieldCurve-class}
\alias{YieldCurve-class}
\alias{YieldCurve}
\title{Yield curves}
\description{
Yield curves
}
\section{Methods}{

\describe{
\item{\code{EuroArea()}}{Query Euro area yield curves published by the ECB}

\item{\code{IndiaZeroCoupon()}}{Query the Indian Zero Coupon Yield Curve published by CCIL}

\item{\code{UsTreasury()}}{Query the Daily Treasury Yield Curve Rates published by US Treasury}
}}

\references{
\url{https://www.ccilindia.com}
\url{http://treasury.gov}
\url{https://www.ecb.europa.eu}

Read the python documentation for information on the data-attributes \url{https://plutopy.readthedocs.io/en/latest/YieldCurve.html}

Sample notebook: \url{https://github.com/shyams80/plutons/blob/master/docs-R/YieldCurve.ipynb}
}
