% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_compare.R
\name{extract_compare}
\alias{extract_compare}
\title{extract the result of \code{geom_compare} from a \code{ggplot} object}
\usage{
extract_compare(p)
}
\arguments{
\item{p}{ggplot object}
}
\value{
compare tibble
}
\description{
extract the result of \code{geom_compare} from a \code{ggplot} object
}
