% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-Fred.R
\docType{class}
\name{Fred-class}
\alias{Fred-class}
\alias{Fred}
\title{St. Louis Fed FRED cache}
\arguments{
\item{SERIES_ID}{integer, use this to query the time-series}

\item{TICKER}{string, FRED id}

\item{NAME}{string, title}

\item{FREQUENCY}{string, 'D' for day, 'M' for month, 'Q' for quarter, etc.}

\item{UNITS}{string, 'Miles', 'Million of Dollars', etc.}

\item{SEASON_ADJUST}{string, 'SA' for seasonally adjusted, etc.}

\item{TIME_STAMP}{date}

\item{VAL}{numeric}
}
\description{
St. Louis Fed FRED cache
}
\section{Methods}{

\describe{
\item{\code{Meta()}}{Query the FRED meta-data cache}

\item{\code{TimeSeries()}}{Query the FRED time-series cache}
}}

\examples{
library(dbplyr)
library(dplyr)
library(odbc)
library(plutoR)
options("scipen"=999)

# get all series with the name 'India ' in it with the first and last dates
# for which time-series data are available
fred<-Fred()
indianFred <- fred$Meta() \%>\% inner_join(fred$TimeSeries(), by = 'SERIES_ID') \%>\%
    filter(NAME \%like\% '\%India \%') \%>\%
    group_by(SERIES_ID, TICKER, NAME) \%>\%
    summarize(MIN_TS = min(TIME_STAMP), MAX_TS = max(TIME_STAMP)) \%>\%
    select(SERIES_ID, TICKER, NAME, MIN_TS, MAX_TS)

print(data.frame(indianFred))

# get a sample of one of the time-series
fred$TimeSeries() \%>\% filter(SERIES_ID == -2147478748, TIME_STAMP >= as.Date('2019-01-01')) \%>\%
    arrange(desc(TIME_STAMP)) \%>\%
    head(5) \%>\%
    print()

}
\references{
\url{https://fred.stlouisfed.org/}
}
