% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_compare.R
\name{geom_compare}
\alias{geom_compare}
\title{add p value and fold change on a plot}
\usage{
geom_compare(
  mapping = NULL,
  data = NULL,
  stat = "compare",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  y_position = NULL,
  step_increase = 0.05,
  tip_length = 0.02,
  cp_label = c("psymbol"),
  ns_lineheight_just = 0.2,
  ignore_ns = FALSE,
  fc_method = NULL,
  comparisons = NULL,
  paired = FALSE,
  alternative = "two.sided",
  test_method = "wilcoxon",
  ns_symbol = "NS",
  cp_ref = NULL,
  cp_inline = FALSE,
  brackets_widen = 0,
  fc_digits = 2,
  cp_result = NULL,
  cp_manual = NULL
)
}
\arguments{
\item{...}{Other arguments passed on to \code{layer()}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{y_position}{y position of the brackets}

\item{step_increase}{the increase height for next bracket,
a ratio according to the whole panel height, default as 0.05}

\item{tip_length}{the length for tips at the ends of the brackets,
a ratio according to the whole panel height, default as 0.02}

\item{cp_label}{which values will be add on the plot, a character vector
with some of \verb{psymbol, p, right_deno_fc, left_deno_fc} in it.
If \code{comparisons} is assigned, you can also include \verb{fc1, fc2}}

\item{ns_lineheight_just}{if show \code{psymbol} in the label, justify
the \code{NS} labels to make the lineheights look balanced}

\item{ignore_ns}{if \code{TRUE} will ignore all label items if p >= 0.05,
or you can assign a character vector like \code{cp_label} to ignore some items
of the label}

\item{fc_method}{fold change method, default is \code{mean}. If you use \code{log10} or
\code{log2} axis, default is \code{geom_mean}.}

\item{comparisons}{a list of two-element vector, to assign the comparisons
should be performed}

\item{paired}{paired test or not, \code{FALSE} as default. If \code{TRUE}, you should
use mapping=aes(paired_by=col) to indicate pairs by an extra column}

\item{alternative}{one of \verb{two.sided, greater, less}}

\item{test_method}{\code{wilcoxon} as default, one of \verb{wilcoxon, t}}

\item{ns_symbol}{the symbol of non-significant, \code{NS} as default}

\item{cp_ref}{reference item, the others will be compared with it}

\item{cp_inline}{draw in line or not, default is \code{FALSE}}

\item{brackets_widen}{widen the brackets, can be a negative value}

\item{fc_digits}{fold change digits}

\item{cp_manual}{manual comparisons table,
please refer to \code{extract_compare()}}
}
\description{
add p value and fold change on a plot
}
