% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-Indices.R
\docType{class}
\name{Indices-class}
\alias{Indices-class}
\alias{Indices}
\title{Indices}
\description{
Indices
}
\section{Methods}{

\describe{
\item{\code{BarclaysMeta()}}{Meta information about the BarclaysTimeSeries by TICKER}

\item{\code{BarclaysTimeSeries()}}{Query the index time-series published by Barclays}

\item{\code{BseConstituents()}}{Query the latest constituents of BSE indices}

\item{\code{BseTimeSeries()}}{Query the index time-series published by the BSE}

\item{\code{IndiaGsecTimeSeries()}}{Query the Indian Government Soverign Bond index time-series published by the CCIL}

\item{\code{IndiaVixTimeSeries()}}{Query the India VIX time-series data published by the NSE}

\item{\code{NseConstituents()}}{Query the latest constituents of NSE indices}

\item{\code{NseTimeSeries()}}{Query the index time-series published by the NSE}

\item{\code{YahooFinanceTimeSeries()}}{Query the index time-series published by Yahoo Finance}
}}

\references{
\url{https://nseindia.com/}
\url{https://bseindia.com/}
\url{https://www.ccilindia.com}
\url{https://finance.yahoo.com/}
\url{https://indices.barclays}

Read the python documentation for information on the data-attributes \url{https://plutopy.readthedocs.io/en/latest/Indices.html}

Sample notebook: \url{https://github.com/shyams80/plutons/blob/master/docs-R/Indices.ipynb}
}
