% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-Indices.R
\docType{class}
\name{Indices-class}
\alias{Indices-class}
\alias{Indices}
\title{Indices}
\description{
Indices
}
\section{Methods}{

\describe{
\item{\code{BseConstituents()}}{Query the latest constituents of BSE indices}

\item{\code{BseTimeSeries()}}{Query the index time-series published by the BSE}

\item{\code{NseConstituents()}}{Query the latest constituents of NSE indices}

\item{\code{NseTimeSeries()}}{Query the index time-series published by the NSE}
}}

\references{
\url{https://nseindia.com/}
\url{https://bseindia.com/}

Read the python documentation for information on the data-attributes \url{https://plutopy.readthedocs.io/en/latest/Indices.html}

Sample notebook: \url{https://github.com/shyams80/plutons/blob/master/docs-R/Indices.ipynb}
}
