% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plutor-package.R
\docType{package}
\name{plutor-package}
\alias{plutor}
\alias{plutor-package}
\title{plutor: Useful Functions for Visualization}
\description{
In ancient Roman mythology, 'Pluto' was the ruler of the underworld and presides over the afterlife. 'Pluto' was frequently conflated with 'Plutus', the god of wealth, because mineral wealth was found underground. When plotting with R, you try once, twice, practice again and again, and finally you get a pretty figure you want. It's a 'plot tour', a tour about repetition and reward. Hope 'plutor' helps you on the tour!
}
\author{
\strong{Maintainer}: William Song \email{william_swl@163.com}

}
\keyword{internal}
