% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_pl}
\alias{theme_pl}
\title{a new extensible theme}
\usage{
theme_pl(
  base_size = 10,
  size_scales = c(5, 5, 6, 8, 10, 10),
  base_lw = 1,
  margin_factor = 0.25,
  font_family = "",
  ...
)
}
\arguments{
\item{base_size}{base size of fonts and margins}

\item{size_scales}{a vector of element size scales, namely:
\enumerate{
\item base size, used by legend text, axis text, caption
\item used by annotation labels
\item used by legend title, axis title, strip text
\item used by subtitle
\item used by title
\item used by tag
}}

\item{base_lw}{base line width}

\item{margin_factor}{the factor of margin to size, default is 0.25}

\item{font_family}{font family}

\item{...}{}
}
\value{
theme object of ggplotusethis::use_version()
}
\description{
a new extensible theme
}
\examples{
theme_pl()
}
