% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{tpt}
\alias{tpt}
\alias{ppt}
\title{trans geom text or point to the real point}
\usage{
tpt(x)

ppt(x)
}
\arguments{
\item{x}{text point in geom}
}
\value{
real point
}
\description{
trans geom text or point to the real point
}
\examples{
tpt(1)
}
